# KR2/fbr-practice_9_10/backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FBR Practice 9 Backend API",
    "version": "1.0.0",
    "description": "Backend API for practices 9 and 10"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-refresh-token"
      }
    },
    "schemas": {
...
      "CreateProductRequest": {
        "type": "object",
        "required": ["title", "category", "description", "price"],
        "properties": {
          "title": {
            "type": "string",
            "example": "Lego"
          },
          "category": {
            "type": "string",
            "example": "Lego Technic"
          },
          "description": {
            "type": "string",
            "example": "Конструктор для детей от 14 лет"
          },
          "price": {
            "type": "number",
            "example": 99999
          }
        }
      },
      "UpdateProductRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Lego Star Wars"
          },
          "category": {
            "type": "string",
            "example": "Lego Technic"
          },
          "description": {
            "type": "string",
            "example": "Конструктор для детей от 10 лет"
          },
          "price": {
            "type": "number",
            "example": 109999
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/products": {
      "post": {
        "summary": "Создать товар",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Товар создан"
          },
          "400": {
            "description": "Некорректные данные"
          }
        }
      },
      "get": {
        "summary": "Получить список товаров",
        "responses": {
          "200": {
            "description": "Список товаров"
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Получить товар по id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Товар найден"
          },
          "401": {
            "description": "Требуется access token"
          },
          "404": {
            "description": "Товар не найден"
          }
        }
      },
      "put": {
        "summary": "Обновить товар",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Товар обновлён"
          },
          "401": {
            "description": "Требуется access token"
          },
          "404": {
            "description": "Товар не найден"
          }
        }
      },
      "delete": {
        "summary": "Удалить товар",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Товар удалён"
          },
          "401": {
            "description": "Требуется access token"
          },
          "404": {
            "description": "Товар не найден"
          }
        }
      }
    }
  }
}